Promiseを繋げるのではなく、awaitで待つように修正
